requirements.txt
customtkinter
pystray
Pillow
comtypes
screeninfo
